+++ app/src/main/AndroidManifest.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.example.projectpopularlibraries">
+    package="com.example.project_popular_libraries">
 
     <application
         android:allowBackup="true"
-        android:name=".app.App"
+        android:name="com.example.project_popular_libraries.app.App"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
@@ -13,8 +13,8 @@
         android:theme="@style/AppTheme"
         tools:ignore="AllowBackup,GoogleAppIndexingWarning">
 
-        <activity android:name=".action_processing_activity.view.ActionProcessingActivity" />
-        <activity android:name=".photo_layout_activity.view.PhotoLayoutActivity">
+        <activity android:name="com.example.project_popular_libraries.action_processing_activity.view.ActionProcessingActivity" />
+        <activity android:name="com.example.project_popular_libraries.photo_layout_activity.view.PhotoLayoutActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
