+++ app/src/main/java/com/example/project_popular_libraries/action_processing_activity/view/IActionView.java
@@ -1,0 +1,7 @@
+package com.example.project_popular_libraries.action_processing_activity.view;
+
+import moxy.MvpView;
+
+public interface IActionView extends MvpView {
+
+}
