+++ app/src/main/java/com/example/project_popular_libraries/photo_layout_activity/view/IRecyclerPhotoViewHolder.java
@@ -1,0 +1,6 @@
+package com.example.project_popular_libraries.photo_layout_activity.view;
+
+public interface IRecyclerPhotoViewHolder {
+    void setImage(String url);
+    int getPos();
+}
