--- app/src/main/java/com/example/projectpopularlibraries/action_processing_activity/presenter/ActionProcessingPresenter.java
@@ -1,10 +1,0 @@
-package com.example.projectpopularlibraries.action_processing_activity.presenter;
-
-
-import com.example.projectpopularlibraries.action_processing_activity.view.IActionView;
-
-import moxy.MvpPresenter;
-
-public class ActionProcessingPresenter extends MvpPresenter<IActionView> {
-    private static final String TAG = "ActionProcessingPresenter";
-}
