--- app/src/main/java/com/example/projectpopularlibraries/action_processing_activity/view/IActionView.java
@@ -1,7 +1,0 @@
-package com.example.projectpopularlibraries.action_processing_activity.view;
-
-import moxy.MvpView;
-
-public interface IActionView extends MvpView {
-
-}
