--- app/src/main/java/com/example/projectpopularlibraries/model/photoupload/entity/Hit.java
@@ -1,25 +1,0 @@
-package com.example.projectpopularlibraries.model.photoupload.entity;
-
-
-import androidx.room.Entity;
-import androidx.room.PrimaryKey;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-@Entity(tableName = "table_hits")
-public class Hit {
-
-    @PrimaryKey(autoGenerate = true)
-    public int id;
-
-    @Expose
-    @SerializedName("webformatURL")
-    public String webformatURL;
-
-    @Override
-    public String toString() {
-        return "Hit{" + "id=" + id +
-                ", webURL = '" + webformatURL + '\'' + '}';
-    }
-}
