--- app/src/main/java/com/example/projectpopularlibraries/model/photoupload/entity/Photo.java
@@ -1,17 +1,0 @@
-package com.example.projectpopularlibraries.model.photoupload.entity;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-import java.util.List;
-
-public class Photo {
-
-    @Expose
-    @SerializedName("totalHits")
-    public int totalHits;
-
-    @Expose
-    @SerializedName("hits")
-    public List<Hit> hits;
-}
