--- app/src/main/java/com/example/projectpopularlibraries/model/photoupload/retrofit/IApiService.java
@@ -1,12 +1,0 @@
-package com.example.projectpopularlibraries.model.photoupload.retrofit;
-
-import com.example.projectpopularlibraries.model.photoupload.entity.Photo;
-
-import io.reactivex.Observable;
-import retrofit2.http.GET;
-import retrofit2.http.Query;
-
-public interface IApiService {
-    @GET("api")
-    Observable<Photo> getPhoto(@Query("key") String key);
-}
