--- app/src/main/java/com/example/projectpopularlibraries/photo_layout_activity/view/IRecyclerPhotoViewHolder.java
@@ -1,6 +1,0 @@
-package com.example.projectpopularlibraries.photo_layout_activity.view;
-
-public interface IRecyclerPhotoViewHolder {
-    void setImage(String url);
-    int getPos();
-}
